+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.4.12 - Unreleased
 
+### Added
+* Add `kwplot.close_figures` and `kwplot.all_figures`.
+
 
 ## Version 0.4.11 - Released 2022-02-16